refactor: 게시글 목록 컴포넌트를 공통 컴포넌트로 분리하여 재사용
