+++ src/containers/mypage/MyBoardListContainer.js
@@ -2,10 +2,8 @@ import MyBoardList from "../../components/mypage/MyBoardList";
 import {useDispatch, useSelector} from "react-redux";
 import {useEffect} from "react";
 import {myBoardList} from "../../modules/mypage";
-import {useNavigate} from "react-router-dom";
 
 const MyBoardListContainer = () => {
-    const navigate = useNavigate();
     const dispatch = useDispatch();
     const {my, myError, loading} = useSelector(({myPage, loading}) => ({
         my: myPage.my,
@@ -17,15 +15,10 @@ const MyBoardListContainer = () => {
         dispatch(myBoardList());
     }, [dispatch]);
 
-    const onNavigateDetail = (boardId) => {
-        navigate(`/boards/${boardId}`);
-    };
-
     return (
         <MyBoardList myBoardList={my}
                      myBoardListError={myError}
                      loading={loading}
-                     onNavigateDetail={onNavigateDetail}
         />
     );
 };
